Move jest client provider to storage module.
