+++ graylog-storage-elasticsearch6/src/main/java/org/graylog/storage/elasticsearch6/jest/JestClientProvider.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.bindings.providers;
+package org.graylog.storage.elasticsearch6.jest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.joschi.jadconfig.util.Duration;
